requirements.txt
langchain
langchain-community
sentence-transformers
faiss-cpu
pandas
streamlit
PyMuPDF
requests
transformers
torch
accelerate
bitsandbytes